Insert text into/over a contentEditable selection
Adds a button that when clicked inserts text into or over a contentEditable
selection. The code was taken from a StackOverflow question/answer (linked to in
the new custom-js.js file). I would like this to insert a span with
mathquill editable, but I don't have time to do that right now. RE: Issue #2