Insert span element in a contentEditable selection
Changes the createTextNode from the last commit to create a span element
in/over the current selection instead. The contentEditable attribute is
also set to false after the insert. Optionally a class can be added to
this but it is currently commented out, because this won't mathquill-ify
the element, a quick attempt to mathquillify is also commented out because it doesn't work. RE: Issue #2